pyproject.toml
[tool.poetry]
name = "python-express"
version = "0.1.0"
description = ""
authors = ["TomHawk123 <matthew.mcmorries@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.1"
pymongo = "^4.6.2"
flask-socketio = "^5.3.6"
flask-restx = "^1.3.0"
gunicorn = "^21.2.0"
flask-talisman = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
